Move dg_1D_vizualizer.py to sfepy/discrete/dg.
