+++ examples/dg/diffusion/example_dg_laplace2D.py
@@ -158,7 +158,7 @@ def define(filename_mesh=None,
         'nls'             : 'newton',
         'ls'              : 'ls',
         'output_format'   : 'msh',
-		'file_format'     : 'gmsh-dg',
+        'file_format'     : 'gmsh-dg',
         # 'pre_process_hook': get_cfl_setup(cfl)
     }
     return locals()
