+++ sfepy/discrete/dg/dg_1D_vizualizer.py
